==> exp/exp07-survival_analyses/03-survival_tests_CNA_insitu.Rmd <==
---
title: "survival CNA"
author: "Noor Rizvi"
date: "2025-02-18"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(tidyr)
library(dplyr)
library(tibble)
library(survival)
library(survminer)
library(purrr)
library(pheatmap)
library(cluster)

```


```{r}

cytoband <- read.table("/path/to/Rdata/output/exp06/cnapp/region_analyses/cytobands/cna_profile_Cytobands_by_ID2024-09-12 10_47_01.867477.tsv", sep="\t", header = TRUE, row.names = 1)
cl <- read.table("/path/to/Rdata/output/exp06/cnapp/res-seg_cna-scores/allsamples_CNApp_CNA_Scores_&_annotations_2024-09-10 10_45_15.475147.tsv", sep="\t", header = TRUE, row.names = 1)


source("../../src/survival.R")
cl$ipsi_lr <- !is.na(cl$hist_1st_lr)
cl <- adjust_survival(cl, cutoff=10)
cl$event.10.chr <- as.character(cl$event.10)

cl$event.10.inv <- ifelse(cl$event.10==1 & cl$hist_1st_lr=="DCIS", 0, cl$event.10)

cl$event.10.insitu <- ifelse(cl$event.10==1 & cl$hist_1st_lr=="INVA", 0, cl$event.10)

cl$event.type.10 <- "no LR"
cl$event.type.10 <- ifelse(cl$event.10.insitu==1, "in-situ LR", cl$event.type.10)
cl$event.type.10 <- ifelse(cl$event.10.inv==1, "invasive LR", cl$event.type.10)




# Format cytoband data
cytoband_long <- cytoband %>%
  rownames_to_column("cytoband") %>%
  pivot_longer(
    cols = -cytoband,
    names_to = "Tumor_Sample_Barcode",
    values_to = "value"
  )

# Gain loss indicators
cytoband_long$gain <- cytoband_long$value>0
cytoband_long$loss <- cytoband_long$value<0

# Merge with clinical data
# Rename the 'index' column if needed
cytoband_long <- left_join(cytoband_long, cl)

```


```{r}

cytoband_long$nugradef2 <- ifelse(cytoband_long$nugradef=="Low", "Moderate", cytoband_long$nugradef)

analyze_cytoband <- function(data, n_permutations = 1000, seed = 1234) {
  cytoband <- unique(data$cytoband)
  
  n_gain <- sum(data$gain, na.rm = TRUE)
  n_loss <- sum(data$loss, na.rm = TRUE)
  
  # Function to calculate permuted p-value
  get_permuted_pvalue <- function(real_model, data, n_perm, alteration_type = "gain") {
    real_coef <- coef(real_model)[1]
    perm_coefs <- numeric(n_perm)
    
    # Get indices of samples with alteration
    if(alteration_type == "gain") {
      altered_indices <- which(data$gain == TRUE)
    } else {
      altered_indices <- which(data$loss == TRUE)
    }
    n_altered <- length(altered_indices)
    
    set.seed(seed)
    
    for(i in 1:n_perm) {
      # Permute while maintaining the same number of alterations
      perm_data <- data
      if(alteration_type == "gain") {
        perm_data$gain <- FALSE  # Reset all to FALSE
        # Randomly select the same number of samples to have alterations
        perm_indices <- sample(1:nrow(data), n_altered)
        perm_data$gain[perm_indices] <- TRUE
      } else {
        perm_data$loss <- FALSE  # Reset all to FALSE
        perm_indices <- sample(1:nrow(data), n_altered)
        perm_data$loss[perm_indices] <- TRUE
      }
      
      # Fit model with permuted data
      tryCatch({
        if(alteration_type == "gain") {
          perm_model <- coxphf(Surv(time.10, event.10.insitu) ~ gain + nugradef2, data = perm_data)
        } else {
          perm_model <- coxphf(Surv(time.10, event.10.insitu) ~ loss + nugradef2, data = perm_data)
        }
        perm_coefs[i] <- coef(perm_model)[1]
      }, error = function(e) {
        perm_coefs[i] <- NA
      })
    }
    
    perm_pvalue <- mean(abs(perm_coefs) >= abs(real_coef), na.rm = TRUE)
    return(perm_pvalue)
  }
  
  # Analysis for gain
  gain_analysis <- if(n_gain >= 5) {tryCatch({
    surv_fit_gain <- survfit(Surv(time.10, event.10.insitu) ~ gain, data = data)
    plot_gain <- ggsurvplot(
      surv_fit_gain,
      data = data,
      pval = TRUE,
      conf.int = TRUE,
      risk.table = TRUE,
      xlab = "Time",
      ylab = "Survival probability",
      title = paste("Kaplan-Meier Curve for", cytoband, "(Gain)")
    )
    cox_model_gain <- coxphf(Surv(time.10, event.10.insitu) ~ gain + nugradef2, data = data)
    perm_pval_gain <- get_permuted_pvalue(cox_model_gain, data, n_permutations, "gain")
    
    list(plot = plot_gain, 
         cox_model = cox_model_gain, 
         permuted_pvalue = perm_pval_gain)
  }, error = function(e) NULL)
  } else NULL
  
  # Analysis for loss
  loss_analysis <- if(n_loss >= 5) {tryCatch({
    surv_fit_loss <- survfit(Surv(time.10, event.10.insitu) ~ loss, data = data)
    plot_loss <- ggsurvplot(
      surv_fit_loss,
      data = data,
      pval = TRUE,
      conf.int = TRUE,
      risk.table = TRUE,
      xlab = "Time",
      ylab = "Survival probability",
      title = paste("Kaplan-Meier Curve for", cytoband, "(Loss)")
    )
    cox_model_loss <- coxphf(Surv(time.10, event.10.insitu) ~ loss + nugradef2, data = data)
    perm_pval_loss <- get_permuted_pvalue(cox_model_loss, data, n_permutations, "loss")
    
    list(plot = plot_loss, 
         cox_model = cox_model_loss, 
         permuted_pvalue = perm_pval_loss)
  }, error = function(e) NULL)
  } else NULL
  
  if(is.null(gain_analysis) && is.null(loss_analysis)) {
    return(NULL)
  }
  
  return(list(
    cytoband = cytoband,
    gain = gain_analysis,
    loss = loss_analysis
  ))
}

results <- cytoband_long %>%
  group_by(cytoband) %>%
  group_map(~analyze_cytoband(.x), .keep = TRUE)


```


```{r}

# Function to safely extract gain and loss p-values from Cox models
safe_extract_info <- function(x) {
  if (is.null(x) || is.null(x$cytoband)) {
    return(NULL)
  }
  
  # Create a single-row data frame with default NA values
  result <- data.frame(
    cytoband = x$cytoband,
    gain_pvalue = NA_real_,
    gain_cox_pvalue = NA_real_,
    gain_coef = NA_real_,
    loss_pvalue = NA_real_,
    loss_cox_pvalue = NA_real_,
    loss_coef = NA_real_,
    stringsAsFactors = FALSE
  )
  
  # Update gain values if they exist
  if (!is.null(x$gain)) {
    result$gain_pvalue <- x$gain$permuted_pvalue
    if (!is.null(x$gain$cox_model)) {
      coefs <- coef(x$gain$cox_model)
      pvals <- x$gain$cox_model$p
      if ("gainTRUE" %in% names(coefs)) {
        result$gain_coef <- coefs["gainTRUE"]
        result$gain_cox_pvalue <- pvals["gainTRUE"]
      }
    }
  }
  
  # Update loss values if they exist
  if (!is.null(x$loss)) {
    result$loss_pvalue <- x$loss$permuted_pvalue
    if (!is.null(x$loss$cox_model)) {
      coefs <- coef(x$loss$cox_model)
      pvals <- x$loss$cox_model$p
      if ("lossTRUE" %in% names(coefs)) {
        result$loss_coef <- coefs["lossTRUE"]
        result$loss_cox_pvalue <- pvals["lossTRUE"]
      }
    }
  }
  
  return(result)
}


# Then just use:
cytoband_pvalues <- map_dfr(results, safe_extract_info)

significant_cytobands_up <- cytoband_pvalues %>%
  filter(
    (!is.na(gain_pvalue) & gain_pvalue < 0.05 & gain_cox_pvalue < 0.05 & gain_coef > 0) |
    (!is.na(loss_pvalue) & loss_pvalue < 0.05 & loss_cox_pvalue < 0.05 & loss_coef > 0)
  )

significant_cytobands_dn <- cytoband_pvalues %>%
  filter(
    (!is.na(gain_pvalue) & gain_pvalue < 0.05 & gain_cox_pvalue < 0.05 & gain_coef < 0) |
    (!is.na(loss_pvalue) & loss_pvalue < 0.05 & gain_cox_pvalue < 0.05 & loss_coef < 0)
  )



# Print significant cytobands
print(significant_cytobands_up)
print(significant_cytobands_dn)

```
